+++ src/main/java/com/launchdarkly/client/TestFeatureStore.java
@@ -1,11 +1,14 @@
 package com.launchdarkly.client;
 
-import com.google.gson.JsonElement;
-import com.google.gson.JsonPrimitive;
+import static com.launchdarkly.client.VersionedDataKind.FEATURES;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
+
+import com.google.gson.JsonElement;
+import com.google.gson.JsonPrimitive;
 
 /**
  * A decorated {@link InMemoryFeatureStore} which provides functionality to create (or override) true or false feature flags for all users.
@@ -13,7 +16,7 @@ import java.util.concurrent.atomic.AtomicInteger;
  * Using this store is useful for testing purposes when you want to have runtime support for turning specific features on or off.
  */
 public class TestFeatureStore extends InMemoryFeatureStore {
-  private static List<JsonElement> TRUE_FALSE_VARIATIONS = Arrays.asList(
+  static List<JsonElement> TRUE_FALSE_VARIATIONS = Arrays.asList(
       (JsonElement) (new JsonPrimitive(true)),
       (JsonElement) (new JsonPrimitive(false))
   );
@@ -34,45 +37,29 @@ public class TestFeatureStore extends InMemoryFeatureStore {
             .variations(TRUE_FALSE_VARIATIONS)
             .version(version.incrementAndGet())
             .build();
-    upsert(key, newFeature);
+    upsert(FEATURES, newFeature);
   }
 
   /**
    * Turns a feature, identified by key, to evaluate to true for every user. If the feature rules already exist in the store then it will override it to be true for every {@link LDUser}.
    * If the feature rule is not currently in the store, it will create one that is true for every {@link LDUser}.
    *
-   * @param key the key of the feature flag to evaluate to true.
+   * @param key the key of the feature flag to evaluate to true
    */
   public void setFeatureTrue(String key) {
     setBooleanValue(key, true);
   }
-
-  /**
-   * @deprecated use {@link #setFeatureTrue(String key)}
-   */
-  @Deprecated
-  public void turnFeatureOn(String key) {
-    setFeatureTrue(key);
-  }
-
+  
   /**
    * Turns a feature, identified by key, to evaluate to false for every user. If the feature rules already exist in the store then it will override it to be false for every {@link LDUser}.
    * If the feature rule is not currently in the store, it will create one that is false for every {@link LDUser}.
    *
-   * @param key the key of the feature flag to evaluate to false.
+   * @param key the key of the feature flag to evaluate to false
    */
   public void setFeatureFalse(String key) {
     setBooleanValue(key, false);
   }
-
-  /**
-   * @deprecated use {@link #setFeatureFalse(String key)}
-   */
-  @Deprecated
-  public void turnFeatureOff(String key) {
-    setFeatureFalse(key);
-  }
-
+  
   /**
    * Sets the value of an integer multivariate feature flag, for all users.
    * @param key the key of the flag
@@ -112,12 +99,12 @@ public class TestFeatureStore extends InMemoryFeatureStore {
             .variations(Arrays.asList(value))
             .version(version.incrementAndGet())
             .build();
-    upsert(key, newFeature);
+    upsert(FEATURES, newFeature);
   }
   
   @Override
-  public void init(java.util.Map<String,FeatureFlag> features) {
-    super.init(features);
+  public void init(Map<VersionedDataKind<?>, Map<String, ? extends VersionedData>> allData) {
+    super.init(allData);
     initializedForTests = true;
   }